Small refactor to home page
