+++ components/parallax/parallaxG.jsx
@@ -1,0 +1,8 @@
+// eslint-disable-next-line unicorn/filename-case
+import ScrollParallax from 'rc-scroll-anim/lib/ScrollParallax';
+import React from 'react';
+
+export default function ParallaxG(props) {
+  /* eslint-disable react/jsx-props-no-spreading */
+  return <ScrollParallax component="g" {...props} />;
+}
